@@ -16,8 +16,7 @@ repositories {
 
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 28
 
     flavorDimensions "versionCode"
     def appName = System.getenv("JOB_NAME") as String ?: "app"
@@ -34,7 +33,7 @@ android {
         multiDexEnabled true
 
         archivesBaseName = "$appName-$versionName.$buildNumber($versionCode)"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     dataBinding {
@@ -107,27 +106,31 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     //noinspection GradleCompatible
-    compile 'com.android.support:appcompat-v7:25.3.1'
+    compile 'androidx.appcompat:appcompat:1.0.0'
     compile project(':sbaseobjects')
     compile 'com.squareup.picasso:picasso:2.5.2'
     compile project(':sdatamanager')
     compile project(':spushnotificationsdroid')
     compile project(':sutils')
     compile project(':Jsoapp')
-    compile 'com.android.support:multidex:1.0.1'
-    compile 'com.android.support:design:25.3.1'
+    compile 'androidx.multidex:multidex:2.0.0'
+    compile 'com.google.android.material:material:1.0.0'
     compile project(path: ':cameraview')
     compile files('libs/com.datecs.api.jar')
     compile 'commons-net:commons-net:3.3'
     compile 'com.flurry.android:analytics:6.4.2'
     compile 'com.github.bumptech.glide:glide:3.7.0'
-    compile 'com.android.support:support-v4:25.3.1'
-    compile 'com.android.support:recyclerview-v7:25.4.0'
+    compile 'androidx.legacy:legacy-support-v4:1.0.0'
+    compile 'androidx.recyclerview:recyclerview:1.0.0'
     implementation('com.crashlytics.sdk.android:crashlytics:2.7.0@aar') {
         transitive = true;
     }
     implementation files('libs/commons-io-2.6.jar')
-    implementation 'com.android.support:cardview-v7:25.4.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+
+    implementation 'com.google.firebase:firebase-core:17.0.0'
 }
+
+apply plugin: 'com.google.gms.google-services'
