@@ -16,7 +16,8 @@
 
 package com.google.android.cameraview;
 
-import android.support.v4.util.ArrayMap;
+
+import androidx.collection.ArrayMap;
 
 import java.util.Set;
 import java.util.SortedSet;
