@@ -4,7 +4,8 @@ import android.content.Intent;
 import android.content.res.Configuration;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.FragmentTransaction;
+import androidx.fragment.app.FragmentTransaction;
+
 import android.util.Log;
 
 import com.sixtemia.sbaseobjects.objects.SFragmentActivity;
