@@ -1,7 +1,7 @@
 package com.sixtemia.sbaseobjects.recycleranimators.base;
 
-import android.support.v4.view.ViewPropertyAnimatorListener;
-import android.support.v7.widget.RecyclerView;
+import androidx.core.view.ViewPropertyAnimatorListener;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 /**
