@@ -1,6 +1,6 @@
 package com.sixtemia.sbaseobjects.recycleranimators.base;
 
-import android.support.v4.view.ViewCompat;
+import androidx.core.view.ViewCompat;
 import android.view.View;
 
 /**
