@@ -11,8 +11,8 @@ import android.graphics.RectF;
 import android.net.Uri;
 import android.os.Environment;
 import android.provider.MediaStore;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import com.sixtemia.sbaseobjects.tools.pickerImatges.ImageBroadcastReceiver;
