@@ -3,8 +3,8 @@ package com.sixtemia.sbaseobjects.tools.pickerImatges;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.sixtemia.sbaseobjects.tools.ImageTools;
 
