@@ -2,7 +2,8 @@ package com.sixtemia.sbaseobjects.tools.pickerImatges;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.FragmentTransaction;
+import androidx.fragment.app.FragmentTransaction;
+
 import android.view.Menu;
 import android.view.MenuItem;
 
