@@ -1,7 +1,7 @@
 package com.sixtemia.sbaseobjects.views;
 
 import android.content.Context;
-import android.support.v7.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 import android.util.Log;
 
 /**
