@@ -1,7 +1,7 @@
 package com.sixtemia.sdatamanager.datamanager;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.text.TextUtils;
 import android.util.Base64;
 import android.util.Log;
