@@ -7,7 +7,7 @@ import java.util.List;
 import android.content.Context;
 import android.graphics.Paint;
 import android.graphics.Typeface;
-import android.support.v4.util.LruCache;
+import androidx.core.util.LruCache;
 import android.text.Spannable;
 import android.text.TextPaint;
 import android.text.style.MetricAffectingSpan;
