@@ -4,9 +4,9 @@ import java.util.ArrayList;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v4.widget.SwipeRefreshLayout.OnRefreshListener;
+import androidx.core.app.Fragment;
+import androidx.core.widget.SwipeRefreshLayout;
+import androidx.core.widget.SwipeRefreshLayout.OnRefreshListener;
 import android.text.Html;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
