@@ -1,7 +1,8 @@
 package com.sixtemia.spushnotifications;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.FragmentTransaction;
+import androidx.fragment.app.FragmentTransaction;
+
 import android.text.TextUtils;
 import android.view.Menu;
 import android.view.MenuItem;
