@@ -2,7 +2,7 @@ package com.sixtemia.gesbluedroid.activities.passosformulari;
 
 import android.content.Context;
 import android.content.Intent;
-import android.databinding.DataBindingUtil;
+import androidx.databinding.DataBindingUtil;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.os.PersistableBundle;
