@@ -2,10 +2,11 @@ package com.sixtemia.spushnotifications.base;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.FragmentTransaction;
 import android.text.TextUtils;
 import android.view.Menu;
 import android.view.MenuItem;
+
+import androidx.fragment.app.FragmentTransaction;
 
 import com.sixtemia.spushnotifications.R;
 import com.sixtemia.spushnotifications.classes.PreferencesSPush;
